requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
Django==3.0.7
django-crispy-forms==1.9.0
django-mathfilters==1.0.0
idna==2.9
lxml==4.5.0
mysqlclient==1.4.6
pycparser==2.20
PyMySQL==0.9.3
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.8
